Reviewed and made some minor updates to our privacy policy for compliance with Europe's General Data Protection Regulation (GDPR).
Customer privacy has always been very important to us, so very little
of our actual policy has changed in this update. We did add some
information about the purposes and legal basis for our processing, as
well as information about the rights of customers who reside in the
European Economic Area (EEA) and Switzerland or in California.

(And as long as we were going over everything with a fine-tooth comb
anyway, we also fixed a few typos, issues with grammar, and
inconsistent capitalization of terms introduced in the first
paragraph.)

We're also adding a General Data Protection Regulation Data Processing
Addendum, which some of our European customers may need in order to
maintain their own compliance with the GDPR.