Review and update our EULA for 2024.
Add a paragraph to the introduction to give this text more
self-contained context.

Format as Markdown, so the text can be formatted to be easier to scan,
while still keeping revisions easy to compare. Minor changes related
to this:

- Added bold and italic formatting where appropriate.
- Updated whitespace where appropriate.
- Updated quotation marks where appropriate.

Remove references to Registration Keys.

Remove limitations regarding distributing copies of our software. The
software must still be licensed by end users, but there's no need to
prohibit distribution of the software itself.

Introduced a clause (in 13(f)) which clarifies that a mutually
executed agreement covering the subject matter of this EULA will
supersede this generic end user license agreement, not the other way
around.

Introduced a clause (in 13(h)) indicating that we won't publicize who
our customers are without their approval. (This is not a policy
change, just a clarification!)

Break up section 7 into several sections:

- 7. Collection and Use of Your Information. This now references our
  more comprehensive privacy policy (which can still be summarized as
  "At Omni, Your Data Is Yours").

- 8. Use of Software Services; Security; Availability. This now
  defines the term "Software Services". But more importantly, it
  introduces expectations regarding notices of scheduled downtime.

- 9, 10. Unchanged other than to updating formatting, capitalization,
  and using better-defined terms.

Broke up the Termination section into enumerated subsections that are
easier to scan and reference.

Clarify some language to more consistently indicate when licenses are
personal and nontransferable, and to indicate that what is offered is
a license, not a right.

Remove references to specific physical distribution mechanisms such as
CD-ROM. (It's been a while since we've distributed our apps on
CD-ROM!) Also, When asking for a refund, you can simply delete our
software rather than "returning" the software.

Define the terms "Subscription" and "Feedback" and reference them as
appropriate. Also define "we" and "our" in exactly the way you
probably already expected they were defined.

Update the spelling of "macOS" to match Apple's current guidelines.

Use Oxford commas when listing multiple items.

Update this document's copyright, and our current physical address.